1.0.3.16: experimental x86-64/darwin suport  * fix sb-posix time structs to match headers on darwin  * comment out mtime sb-posix test as this isn't working ATM  * add UD2A trap stuff to x86-64 and corresponding word-imm support  * remove bogus extern-alien-name in boxed_region fixups  * add compiler parameters for Darwin  * add x86-64 darwin config file  * some type safety fixes (proper types) in darwin-dlshim.c  * use setpgid on Darwin  * add signal context support x86-64/darwin  * report trap instead of si_code for trap_Error/trap_Cerror  * unsigned -> unsigned long in purify.c  * add mach exception handler support for x86-64/darwin  * x86-64 assembly hacks to make darwin's assembler happy  * update x86-64-bsd-os.c to suppot darwin and mach exceptions  * add x86-64-darwin-os.c/h  * update LDSO stubs for x86-64 darwin

+++ src/runtime/darwin-os.c
@@ -28,7 +28,7 @@ char *
 os_get_runtime_executable_path()
 {
     char path[PATH_MAX + 1];
-    uint32_t size = sizeof(path);
+    size_t size = sizeof(path);
 
     if (_NSGetExecutablePath(path, &size) == -1)
         return NULL;
